added date histogram, plus string query bucketing code
